fix build

+++ beacon-chain/rpc/beaconv1/state_test.go
@@ -171,7 +171,7 @@ func TestGetStateRoot(t *testing.T) {
 		// We fill state roots with hex representations of natural numbers starting with 1.
 		// Example: 16 becomes 0x00...0f
 		fillStateRoots := func(state *pb.BeaconState) {
-			rootsLen := params.MainnetConfig().SlotsPerHistoricalRoot
+			rootsLen := uint64(params.MainnetConfig().SlotsPerHistoricalRoot)
 			roots := make([][]byte, rootsLen)
 			for i := uint64(0); i < rootsLen; i++ {
 				roots[i] = make([]byte, 32)
@@ -255,7 +255,7 @@ func TestGetStateRoot(t *testing.T) {
 			ChainInfoFetcher: chainService,
 		}
 		_, err = s.GetStateRoot(ctx, &ethpb.StateRequest{
-			StateId: []byte(strconv.FormatUint(state.Slot()+1, 10)),
+			StateId: []byte(strconv.FormatUint(uint64(state.Slot()+1), 10)),
 		})
 		assert.ErrorContains(t, "Slot cannot be in the future", err)
 	})
